today we have cooked. the scene stuff is done, viseme stuff is done, now we just need to tie everything together. We are still working on getting the first pass going. Tomorrow or tonight going to hook up the viseme generator, with the TTS and video stitcher, and need to update the containers with scheduling for the news aggregator.
But yeah next session need to finalize the first pass, then setup the scheduling, and containers. Then we can really start cooking. That will probably take the entire session. But after that we need to setup twitch interactivity probably, and probably a bunch of other stuff.

I'm going to have to make the TTS voice better, and make more expressions and visemes for the avatar, add urgency and prioritization guessers, and just update the container generation and scheduling