almost ready for a first pass test, but need to add the blender download
after this we just have to add the ffmpeg queue and we should be good to go for a very rough version that we can deploy. After that we should work on: scheduling the scraper, adding visemes, adding twitch functionality, adding more expressions, and putting intake systems in place for story prioritization. We can also add different scenes and camera angles.